Rename wait pkg to retry
